Rename adminCodes and volunteerHours router files to use camelCase
