+++ app.js
@@ -4,12 +4,12 @@ const cookieParser = require('cookie-parser');
 
 const userRouter = require('./routes/users/users');
 const eventRouter = require('./routes/events/events');
-const volunteerHoursRouter = require('./routes/volunteers/volunteer_hours');
+const volunteerHoursRouter = require('./routes/volunteers/volunteerHours');
 const volunteerRouter = require('./routes/volunteers/volunteers');
 const availabilityRouter = require('./routes/users/availability');
 const waiversRouter = require('./routes/events/waivers');
 const s3UploadRouter = require('./routes/s3Upload');
-const adminCodeRouter = require('./routes/users/admin_codes');
+const adminCodeRouter = require('./routes/users/adminCodes');
 const { authRouter } = require('./routes/auth');
 
 const app = express();
